fix(http3/connect-udp): parse context id as varint (#3018)

+++ neqo-http3/src/features/extended_connect/session.rs
@@ -551,5 +551,5 @@ pub(crate) enum DgramContextIdError {
     #[error("Missing context identifier")]
     MissingIdentifier,
     #[error("Unknown context identifier: {0}")]
-    UnknownIdentifier(u8),
+    UnknownIdentifier(u64),
 }
